fix warning from gcc
git-svn-id: https://iptux.googlecode.com/svn/trunk@173 38dc79a6-912d-11dd-8943-ed8a20f67c6a

+++ src/utils.cpp
@@ -187,7 +187,7 @@ char *assert_filename_inexist(const char *path)
         count = 1;
         while (count) {
                 snprintf(buf, MAX_PATHLEN, "%.*s%" PRIu16 "_%s",
-                                 ptr - path, path, count, ptr);
+                                 (int)(ptr - path), path, count, ptr);
                 if (access(buf, F_OK) != 0)
                         break;
                 count++;
